refactor(services): dedupe seam types and drop port-optional chaining
The seam file is now the single documented home for ArchiveWorkspaceOptions,
SendMessageInternalOptions, and WorkspaceLiveActivity instead of duplicating
workspaceService declarations. StreamErrorRecoveryOutcome comes from its
canonical agentSession export. The AgentTaskStatus re-export shim is gone;
importers use the seam. Method-level optional chaining on both typed ports is
removed: the interfaces guarantee the methods, so the narrow-mock hedges and
their comments no longer apply.
